+++ src/widgets/Badge/Badge.tsx
@@ -7,11 +7,13 @@ export interface BadgeProps {
     children?: ReactNode;
 
     color?: "red" | "teal";
+
+    padding?: "badge" | "button";
 }
 
-export function Badge({ children, color = "red" }: BadgeProps) {
+export function Badge({ children, padding = "badge", color = "red" }: BadgeProps) {
 
-    const css = [ "badge", `badge-${color}` ];
+    const css = [ "badge", `badge-${color}`, `badge-padding-${padding}` ];
 
     return (
         <span className={css.join(" ")}>